ini: preserve EOF semantics when not optimizing
I saw a comment on HN claiming that the lack of buffering via an int
means that the code won't work on ARM.
This was surprising to me, given I exclusively tested this code on ARM.
What the comment was missing is that since you save the character to
a buffer, but then immediately read it, means that the character will
simply not leave the register it got put into post-optimization, meaning
it maintains full semantics.
This, of course, is not "legally C", so it's better off like this.
+ I forgot to make wss static.